# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "葫芦家商城v2",
        "title": "葫芦家",
        "contact": {
            "name": "flaravel"
        },
        "license": {},
        "version": "2.0"
    },
    "host": "localhost:8100",
    "paths": {
        "/api/admin/user": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "\"\"",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台"
                ],
                "summary": "创建管理员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "手机号",
                        "name": "phone",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "description": "角色",
                        "name": "roles",
                        "in": "body",
                        "schema": {
                            "type": "array"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseMessage"
                        }
                    },
                    "401": {
                        "description": "Token解析失败",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseMessage"
                        }
                    },
                    "422": {
                        "description": "请求参数缺失或者不正确返回",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseMessage"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.ResponseMessage": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}